requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
pydantic==2.5.1
passlib[argon2]==1.7.4
pyjwt==2.10.1
aiosmtplib==4.0.2
python-dotenv==1.1.1